+++ modules/apps/headless/headless-workflow/headless-workflow-impl/src/main/java/com/liferay/headless/workflow/internal/resource/WorkflowLogResourceImpl.java
@@ -1,0 +1,57 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.headless.workflow.internal.resource;
+
+import com.liferay.headless.workflow.dto.WorkflowLog;
+import com.liferay.headless.workflow.dto.WorkflowLogCollection;
+import com.liferay.headless.workflow.resource.WorkflowLogResource;
+import com.liferay.portal.vulcan.context.Pagination;
+
+import java.util.Collections;
+
+import javax.annotation.Generated;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ServiceScope;
+import org.osgi.service.jaxrs.whiteboard.JaxrsWhiteboardConstants;
+
+/**
+ * @author Javier Gamarra
+ * @generated
+ */
+@Component(
+	property = {
+		JaxrsWhiteboardConstants.JAX_RS_APPLICATION_SELECT + "=(osgi.jaxrs.name=headless-workflow-application.rest)",
+		JaxrsWhiteboardConstants.JAX_RS_RESOURCE + "=true", "api.version=1.0.0"
+	},
+	scope = ServiceScope.PROTOTYPE, service = WorkflowLogResource.class
+)
+@Generated("")
+public class WorkflowLogResourceImpl implements WorkflowLogResource {
+
+	@Override
+	public WorkflowLog getWorkflowLogs(Integer id) throws Exception {
+		return new WorkflowLog();
+	}
+
+	@Override
+	public WorkflowLogCollection<WorkflowLog> getWorkflowTasksWorkflowLogs(
+			Integer parentId, Pagination pagination)
+		throws Exception {
+
+		return new WorkflowLogCollection(Collections.emptyList(), 0);
+	}
+
+}
