+++ modules/apps/headless/headless-workflow/headless-workflow-impl/src/main/java/com/liferay/headless/workflow/internal/resource/WorkflowTaskResourceImpl.java
@@ -14,11 +14,9 @@
 
 package com.liferay.headless.workflow.internal.resource;
 
-import com.liferay.headless.workflow.dto.WorkflowLog;
-import com.liferay.headless.workflow.dto.WorkflowLogCollection;
 import com.liferay.headless.workflow.dto.WorkflowTask;
 import com.liferay.headless.workflow.dto.WorkflowTaskCollection;
-import com.liferay.headless.workflow.resource.HeadlessWorkflowResource;
+import com.liferay.headless.workflow.resource.WorkflowTaskResource;
 import com.liferay.portal.vulcan.context.Pagination;
 
 import java.util.Collections;
@@ -38,10 +36,10 @@ import org.osgi.service.jaxrs.whiteboard.JaxrsWhiteboardConstants;
 		JaxrsWhiteboardConstants.JAX_RS_APPLICATION_SELECT + "=(osgi.jaxrs.name=headless-workflow-application.rest)",
 		JaxrsWhiteboardConstants.JAX_RS_RESOURCE + "=true", "api.version=1.0.0"
 	},
-	scope = ServiceScope.PROTOTYPE, service = HeadlessWorkflowResource.class
+	scope = ServiceScope.PROTOTYPE, service = WorkflowTaskResource.class
 )
 @Generated("")
-public class HeadlessWorkflowResourceImpl implements HeadlessWorkflowResource {
+public class WorkflowTaskResourceImpl implements WorkflowTaskResource {
 
 	@Override
 	public WorkflowTaskCollection<WorkflowTask> getRolesWorkflowTasks(
@@ -49,11 +47,6 @@ public class HeadlessWorkflowResourceImpl implements HeadlessWorkflowResource {
 		throws Exception {
 
 		return new WorkflowTaskCollection(Collections.emptyList(), 0);
-	}
-
-	@Override
-	public WorkflowLog getWorkflowLogs(Integer id) throws Exception {
-		return new WorkflowLog();
 	}
 
 	@Override
@@ -67,14 +60,6 @@ public class HeadlessWorkflowResourceImpl implements HeadlessWorkflowResource {
 		throws Exception {
 
 		return new WorkflowTaskCollection(Collections.emptyList(), 0);
-	}
-
-	@Override
-	public WorkflowLogCollection<WorkflowLog> getWorkflowTasksWorkflowLogs(
-			Integer parentId, Pagination pagination)
-		throws Exception {
-
-		return new WorkflowLogCollection(Collections.emptyList(), 0);
 	}
 
 	@Override
